Fixing all bugs

+++ app/Http/Controllers/ListingController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use App\Models\Listing;
 use Faker\Provider\Lorem;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
 
 class ListingController extends Controller
 {
@@ -33,16 +35,25 @@ class ListingController extends Controller
             'website' => 'required',
             'description' => 'required'
          ]);
-   if($request->hasFile('logo')){
-    $formFields['logo'] = $request->file('logo')->store('logos', 'public');
-   }
+        if($request->hasFile('logo')){
+            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
+        }
+        $formFields['user_id'] = auth()->id();
+
+        // dd($formFields);
          Listing::create($formFields);
-         return redirect('/listings')->with('welcome', 'Listing created succesfully');
+         return redirect('/')->with('message', 'Listing created succesfully');
     }
+
     public function edit(Listing $listing){
-        return view('listings.edit', ['listing'=> $listing]) ; 
+        return view('listings.edit', ['listing' => $listing] );
     }
     public function update(Request $request, Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
           $updateFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -53,13 +64,22 @@ class ListingController extends Controller
             'description' => 'required'
          ]);
    if($request->hasFile('logo')){
-    $updateFields['logo'] = $request->file('logo')->store('logos', 'public');
+        $updateFields['logo'] = $request->file('logo')->store('logos', 'public');
     }
    $listing->update($updateFields);
-   return back()->with('welcome', 'Listing updates succefully');
+   return back()->with('message', 'Listing updated succefully');
 }
 public function destroy(Listing $listing){
+    if($listing->id != auth()->id()){
+        abort(403, 'Unauthorized Action');
+    }
+
     $listing->delete();
-    return redirect('/listings')->with('welcome', 'Listing deleted successfully');
+    return redirect('/')->with('message', 'Listing deleted successfully');
+}
+public function manage(){
+    // dd(auth()->id());
+    $listings = Listing::latest()->where('user_id', auth()->id())->get();
+    return view('listings.manage',['listings'=> $listings]);
 }
 }
